fixture file renamed
